Move pkg/kubectl/apps to staging
